Fixed up debug compilation errors.

+++ src/UQMGame.cpp
@@ -450,7 +450,8 @@ uqstl::pair<int, bool> UQMGame::setup(uqstl::span<uqgsl::zstring> args)
 	   ProcessThreadLifecycles... */
 	// initAudio (snddriver, soundflags);
 	// Make sure that the compiler treats multidim arrays the way we expect
-	assert(sizeof(int[ControlTemplate::NUM * NUM_KEYS]) == sizeof(int[static_cast<int>(ControlTemplate::NUM)][NUM_KEYS]));
+	assert(sizeof(int[static_cast<int>(ControlTemplate::NUM) * NUM_KEYS]) == sizeof(int[static_cast<int>(ControlTemplate::NUM)][NUM_KEYS]));
+	
 	TFB_SetInputVectors((int*)ImmediateInputState.menu, NUM_MENU_KEYS,
 						(int*)ImmediateInputState.key, NUM_KEYS,
 						static_cast<int>(ControlTemplate::NUM));
